requirements.txt
python-telegram-bot==20.7
openai==1.12.0
supabase==2.3.4
python-dotenv==1.0.0
pydub==0.25.1
requests==2.31.0
pillow==10.2.0
